backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
stripe==11.4.1
qrcode[pil]==8.0
celery[redis]==5.4.0
resend==2.7.0
boto3==1.36.2
google-auth==2.37.0
google-auth-oauthlib==1.2.1
